requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
arrow==1.4.0
    # via duck-punch-mcp (pyproject.toml)
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
cachetools==6.2.4
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
chardet==5.2.0
    # via duck-punch-mcp (pyproject.toml)
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer
    #   uvicorn
cryptography==46.0.3
    # via
    #   duck-punch-mcp (pyproject.toml)
    #   pyjwt
    #   pyopenssl
google-auth==2.45.0
    # via duck-punch-mcp (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
mcp==1.25.0
    # via duck-punch-mcp (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
protobuf==6.33.2
    # via duck-punch-mcp (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via mcp
pyopenssl==25.3.0
    # via duck-punch-mcp (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   duck-punch-mcp (pyproject.toml)
    #   arrow
python-dotenv==1.2.1
    # via
    #   duck-punch-mcp (pyproject.toml)
    #   mcp
    #   pydantic-settings
python-multipart==0.0.21
    # via mcp
pytz==2025.2
    # via duck-punch-mcp (pyproject.toml)
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   duck-punch-mcp (pyproject.toml)
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via duck-punch-mcp (pyproject.toml)
rich==14.2.0
    # via typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   duck-punch-mcp (pyproject.toml)
    #   python-dateutil
sse-starlette==3.1.2
    # via mcp
starlette==0.50.0
    # via
    #   mcp
    #   sse-starlette
typer==0.21.0
    # via mcp
typing-extensions==4.15.0
    # via
    #   duck-punch-mcp (pyproject.toml)
    #   anyio
    #   mcp
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   referencing
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via arrow
urllib3==2.6.2
    # via requests
uvicorn==0.40.0
    # via mcp
